update source 1.fix Byte conflict defination in zlib and Byte.hpp 2.add Smtp reference doc 3.add destructor in WebSocketLinker/Listener/Server for build error(in instantiation of member function requested here)

+++ lang/include/Byte.hpp
@@ -11,7 +11,7 @@ class _String;
 
 #define byte uint8_t
 
-DECLARE_CLASS(Byte) IMPLEMENTS(Number)<uint16_t> {
+DECLARE_CLASS(Byte) IMPLEMENTS(Number)<uint8_t> {
 public:
     static const int __isReflected = 1;
